pyproject.toml
[tool.poetry]
name = "economic-indicator"
version = "0.1.0"
description = ""
authors = ["Saipraneeth <2506664+msaipraneeth@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.1"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
streamlit = "^1.29.0"
plotly = "^5.18.0"
sqlalchemy = "^2.0.23"
scikit-learn = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
